remove lurker monster for the time being
The Lurker monsters, after recent modifications, are pretty much the same as spiders as they chase and bite the player. Their original purpose of hiding, running from the player and rewarding health is temporarily scrapped.
As this wasn't enough, the model the monster is using is unlicensed: it's just not worth it.
Gonna remove the monster for the time being.